move maintenance project. note does not build!
