+++ DbAppSettings/Source/DbAppSettings.Test/Properties/DbAppSettingsTestSettings.cs
@@ -1,6 +1,6 @@
 ﻿using DbAppSettings.Model.Domain;
 
-namespace DbAppSettings.Test.Mock
+namespace DbAppSettings.Test.Properties
 {
     public class DbAppSettingsTestSettings
     {
