+++ DbAppSettings/Source/WebDbAppSettingsMaintenance/Areas/Maintenance/Controllers/MaintenanceController.cs
@@ -2,11 +2,10 @@
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web.Mvc;
-using DbAppSettings.Model.DataAccess.Maintenance.Demo;
+using DbAppSettings.Maintenance.Model.Service;
+using DbAppSettings.Maintenance.Model.Service.Interfaces;
 using DbAppSettings.Model.DataTransfer;
 using DbAppSettings.Model.Service.Interfaces;
-using DbAppSettings.Model.Service.Maintenance;
-using DbAppSettings.Model.Service.Maintenance.Interfaces;
 using WebDbAppSettingsMaintenance.Areas.Maintenance.Models;
 
 namespace WebDbAppSettingsMaintenance.Areas.Maintenance.Controllers
