+++ DbAppSettings/Source/WebDbAppSettingsMaintenance/packages.config
@@ -4,7 +4,7 @@
   <package id="bootstrap" version="3.0.0" targetFramework="net461" />
   <package id="jQuery" version="1.10.2" targetFramework="net461" />
   <package id="jQuery.Validation" version="1.11.1" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights" version="2.2.0" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights" version="2.2.0" targetFramework="net461" requireReinstallation="true" />
   <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.0.6" targetFramework="net461" />
   <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.2.0" targetFramework="net461" />
   <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.2.0" targetFramework="net461" />
